+++ src/Core/Application/Models/DTOs/Assets/AssetRelationDto.cs
@@ -7,7 +7,7 @@ namespace Application.Models.DTOs.Assets
             
         }
 
-        public AssetRelationDto(int assetId, int count, decimal purchaseValue, decimal currentValue, DateOnly purchaseDate)
+        public AssetRelationDto(int assetId, decimal count, decimal purchaseValue, decimal currentValue, DateOnly purchaseDate)
         {
             AssetId = assetId;
             Count = count;
@@ -17,7 +17,7 @@ namespace Application.Models.DTOs.Assets
         }
 
         public int AssetId { get; init; }
-        public int Count { get; init; }
+        public decimal Count { get; init; }
         public decimal PurchaseValue { get; init; }
         public decimal CurrentValue { get; init; }
         public DateOnly PurchaseDate { get; init; }
