+++ src/Presentation/Client/Models/Assets/AssetForUpdateResponse.cs
@@ -7,7 +7,7 @@ namespace Client.Models.Assets
 
         }
 
-        public AssetForUpdateResponse(int assetId, int count, decimal purchasePrice, decimal salePrice, DateOnly purchaseDate)
+        public AssetForUpdateResponse(int assetId, decimal count, decimal purchasePrice, decimal salePrice, DateOnly purchaseDate)
         {
             AssetId = assetId;
             Count = count;
@@ -17,7 +17,7 @@ namespace Client.Models.Assets
         }
 
         public int AssetId { get; init; }
-        public int Count { get; init; }
+        public decimal Count { get; init; }
         public decimal PurchasePrice { get; init; }
         public decimal SalePrice { get; init; }
         public DateOnly PurchaseDate { get; init; }
