+++ src/Presentation/Client/Models/Assets/UserAssetsForOperationResponse.cs
@@ -6,7 +6,7 @@ namespace Client.Models.Assets
         {
             
         }
-        public UserAssetsForOperationResponse(int currencyId, string title, string? subTitle, int count, decimal purchasePrice, decimal salePrice)
+        public UserAssetsForOperationResponse(int currencyId, string title, string? subTitle, decimal count, decimal purchasePrice, decimal salePrice)
         {
             CurrencyId = currencyId;
             Title = title;
@@ -19,7 +19,7 @@ namespace Client.Models.Assets
         public int CurrencyId { get; init; }
         public string Title { get; init; } = null!;
         public string? SubTitle { get; init; }
-        public int Count { get; init; }
+        public decimal Count { get; init; }
         public decimal PurchasePrice { get; init; }
         public decimal SalePrice { get; init; }
     }
